Backend: Extract LineDrawer and QuadDrawer from RenderUtils
